Refactor - separation complete
